requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pytesseract==0.3.9
opencv-python==4.6.0.66
numpy==1.24.3
pdf2image==1.16.3
python-docx==0.8.11
openai==0.28.0
scikit-learn==1.3.0
pydantic==1.10.7
